Move session graphQL code into sessions module
