+++ .changeset/plenty-rivers-type.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/keystone': patch
+---
+
+Moved session handling GraphQL schema code into the sessions module.
